requirements.txt
langchain
langchain-groq
python-dotenv
pydantic
requests
pyyaml
